+++ 2 WEEK/SON/yebin/src/main/java/com/yebin/yebin/order/Order.java
@@ -4,20 +4,20 @@ public class Order {
     private Long memberId;
     private String itemName;
     private int itemPrice;
-    private int discountPrice;
+    private float discountPrice;
 
-    public Order(Long memberId, String itemName, int itemPrice, int discountPrice){
+    public Order(Long memberId, String itemName, int itemPrice, float discountPrice){
         this.memberId = memberId;
         this.itemName = itemName;
         this.itemPrice = itemPrice;
         this.discountPrice = discountPrice;
     }
 
-    public int calculatePrice(){
+    public float calculatePrice(){
         return itemPrice - discountPrice;
     }
 
-    public int getDiscountPrice(){
+    public float getDiscountPrice(){
         return discountPrice;
     }
